What have I done
Basically, instead of one command being one command,
how a command is handled can now change based
on the number of parameters that are given when the
command is used.

For example, `p!threads` and `p!threads <server>` can
behave like seperate commands... but they're both
invoked using the `threads` command. It's like overloads,
but for bot commands.